requirements.txt
websocket-client
tk
scikit-learn
